Adds sanity check for the term type inside the command manager (#1315)
# Description

- The purpose of this change is to enhance the robustness of the code by
adding a sanity check to ensure that the command term is of a valid type
before it is processed (Just like ActionTerm). This validation step is
crucial for preventing unexpected errors and maintaining system
stability, especially when dealing with dynamically configured command
terms that may vary based on external inputs or configurations. By
validating the type of the command term, we can provide early feedback
to developers and users, helping to catch misconfigurations or type
mismatches during the initial stages. This change aims to improve the
overall reliability and maintainability of the codebase.

- It also fixes iteration over group_cfg_items inside the observation
manager.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>
Co-authored-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>